# docs/swagger.yaml
basePath: /
definitions:
...
  genproto.Post:
    properties:
      Body:
        type: string
      CategoryId:
        type: string
      PostId:
        type: string
      Title:
        type: string
      UserId:
        type: string
    type: object
  genproto.PostModel:
    properties:
      Body:
        type: string
      CategoryId:
        type: string
      Title:
        type: string
      UserId:
        type: string
    type: object
  genproto.PostTag:
    properties:
      PostId:
        type: string
      PostTagId:
        type: string
      TagId:
        type: string
    type: object
  genproto.PostTagModel:
    properties:
      PostId:
        type: string
      TagId:
        type: string
    type: object
  genproto.PostTags:
    properties:
      PostTags:
        items:
          $ref: '#/definitions/genproto.PostTag'
        type: array
    type: object
  genproto.Posts:
    properties:
      Posts:
        items:
          $ref: '#/definitions/genproto.Post'
        type: array
    type: object
...
info:
  contact: {}
  title: Online Forum API Documentation
paths:
  /admin/post/{id}:
    delete:
      consumes:
      - application/json
      description: This api Deletes Post
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: post deleted successfully!
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Deletes Post
      tags:
      - Post
    put:
      consumes:
      - application/json
      description: This api Updates Post
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      - description: Post Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.PostModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Post'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Updates Post
      tags:
      - Post
...
  /post-tag/{id}:
    delete:
      consumes:
      - application/json
      description: This api Deletes PostTag
      parameters:
      - description: Post-Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post-Tag Deleted Successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Deletes PostTag
      tags:
      - Post-Tag
    get:
      consumes:
      - application/json
      description: This api Gets PostTag by id
      parameters:
      - description: Post-Tag ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PostTag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets PostTag by id
      tags:
      - Post-Tag
    put:
      consumes:
      - application/json
      description: This api updates PostTag
      parameters:
      - description: Post-Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Post-Tag data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.PostTagModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PostTag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Updates PostTag
      tags:
      - Post-Tag
  /post-tag/all:
    get:
      consumes:
      - application/json
      description: This api Gets all PostTags
      parameters:
      - description: Post ID
        in: query
        name: post_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PostTags'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets all PostTags
      tags:
      - Post-Tag
  /post-tag/create:
    post:
      consumes:
      - application/json
      description: This api Creates PostTag
      parameters:
      - description: Post Tag Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.PostTagModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.PostTag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Creates PostTag
      tags:
      - Post-Tag
  /post/{id}:
    get:
      consumes:
      - application/json
      description: This api Gets Post by id
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Post'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Gets Post
      tags:
      - Post
  /post/{id}/comments:
    get:
      consumes:
      - application/json
      description: This api Search Comments by Post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Posts'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Search Comments by Post
      tags:
      - Post
  /post/all:
    get:
      consumes:
      - application/json
      description: This api gets all Post
      parameters:
      - description: User Id
        in: query
        name: user_id
        type: string
      - description: Post Title
        in: query
        name: title
        type: string
      - description: Category Id
        in: query
        name: category_id
        type: string
      - description: Post Body
        in: query
        name: body
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Posts'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Gets all Post
      tags:
      - Post
  /post/create:
    post:
      consumes:
      - application/json
      description: This api create Post
      parameters:
      - description: Post Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.PostModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Post'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Creates Post
      tags:
      - Post
  /post/search:
    get:
      consumes:
      - application/json
      description: This api Searchs Post
      parameters:
      - description: Post Title
        in: query
        name: title
        type: string
      - description: Post Body
        in: query
        name: body
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Posts'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Search Post
      tags:
      - Post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"